Start using Composer to manage dependencies.
Git submodules are still available since we don't want to force Composer on
users yet (it is still in development after all), but we have moved them in the
"vendor" subdirectory and configured them to match the same layout produced by
Composer to use a common logic.

We have also added an autoload.php file in the root directory of the repository
that is able to switch between dependencies fetched with Composer or by using
Git submodules. This file is used by both examples and the test suite.