Partially rewrite the current cURL-based HTTP client.
The HTTP client is now aware of session cookies returned by OrientDB in order
to identify an authenticated session.

Now instead of telling the client to reuse or not the underlying resource, we
can decide if each request should be treated as a whole new session. Starting
a new session resets any cookie previously returned by OrientDB, including the
session cookie.