Add two utility scripts to manage our dependencies.
The first one is used to initialize our dependencies using Composer or Git
submodules. Composer is the preferred method when no arguments are passed
to the script, otherwise users can decide which method to use:

  $ ./bin/initialize-dependencies.sh composer
  $ ./bin/initialize-dependencies.sh submodules

The second script is intended to be used to safely clean up the dependency
tree without breaking the directory structure that is needed to support Git
submodules.