Fix parsing of DATETIME values when casting to DateTime.
OrientDB returns DATETIME values as strings with microsecond resolution using
a colon (:) to separate the microseconds part. PHP, on the other hand, expects
the dot character (.) for the microseconds part thus breaking our caster. As a
workaround we try to normalize this difference before passing the date string
to the constructor of the DateTime class.