Add more Composer-related files to bin/clean-dependencies.sh.
Due to recent changes in Composer, all the files related to autoloading
have been moved from vendor/.composer/ to vendor/ so we need to delete
them too when cleaning up dependencies leftovers.

We still use the autogenerated autoload.php file in vendor/.composer/
for now since it is still available for backwards compatibility, but
ideally we should move to the one placed in vendor/ as soon as the new
directory layout is finalized.