fixed a bug in Dijkstra's Algorithm where the nesting may become too much or the algorithm becomes highly inefficient due to full looping through the entire graph's vertices.
the fix here is to only check for connected vertices and stand-alone vertices in the graph are ignored by this implementation.

sorting is also done on the vertex's connections to find the next nearest adjacent ID.

Since we are getting the nearest adjacent vertex through the connections() array, we will remove the getMostAdjacentConnectionId(); method.

Signed-off-by: Sam-Mauris Yong <mauris@hotmail.sg>