Properly handle integer, float, boolean and NULL values in VALUES.
This change is needed to make it possible to insert non-string values
for the fields specified in an INSERT query. This query:

  $query->insert()
        ->into("Test")
        ->fields(array('a', 'b', 'c', 'd', 'e'))
        ->values(array('0', 0, NULL, true, false));

Is now translated into the following SQL code:

  INSERT INTO Test (a, b, c, d, e) VALUES ("0", 0, NULL, TRUE, FALSE)

In addition to that, values equal to FALSE were being stripped using
array_filter() which is not always a desirable behavior. The formatter
used for LIMIT had to be fixed to accomodate this change.