Remove Doctrine\OrientDB\Formatter\InflectorInterface.
This interface should really be in doctrine/common, but anyway our
usage in Doctrine\ODM\OrientDB\Mapper is dirty at best since we use
it in its constructor signature but fall back to a class that does
not implement it when no custom inflector is specified.

In the end it's definitely better to remove it waiting for a better
solution, we can always extend Doctrine\Common\Util\Inflector if
we need to customize the inflection rules.