Reorganize the Doctrine\OrientDB\Query namespace and related classes.
We basically moved there a few additional classes and interfaces in
addition to the following namespaces:

  - Doctrine\OrientDB\Validator
  - Doctrine\OrientDB\Formatter

in order to make it possible to split the query builder into a separate
subproject in the future.

We should really tweak the naming of certain classes and sub-namespaces,
but the major work is basically done.