Reorganize structure of the test suite.
This change is needed to be able to split the library in two different
repositories (one for the access abstraction layer and one for the ODM).

Integration tests are now enabled by default since it is now easy enough
to bootstrap an instance of OrientDB in a matter of seconds with our CI
scripts, but they can be disabled by removing the XML comment tags from
the "integration" node in the group exclusion list stored in phpunit.xml.