Add the ability to specify query language.
It is now possible to optionally specify which query language to
use when sending queries to OrientDB using HttpBinding::query()
and HttpBinding::command().

This commit introduces a breaking change in the public binding
API since since we had to add a new $language parameter to both
the above mentioned methods and move $database as the last one.